IHaskell.Eval.Parser

Documentation

parseString :: GhcMonad m => String -> m [Located CodeBlock]

Parse a string into code blocks.

data CodeBlock

A block of code to be evaluated. Each block contains a single element - one declaration, statement, expression, etc. If parsing of the block failed, the block is instead a ParseError, which has the error location and error message.

Constructors

Expression String
A Haskell expression.

Declaration String
A data type or function declaration.

Statement String
A Haskell statement (as if in a `do` block).

Import String
An import statement.

TypeSignature String
A lonely type signature (not above a function declaration).

Directive DirectiveType String
An IHaskell directive.

Module String
A full Haskell module, to be compiled and loaded.

ParseError StringLoc ErrMsg
An error indicating that parsing the code block failed.

data StringLoc

Constructors

Loc LineNumber ColumnNumber

data DirectiveType

Directive types. Each directive is associated with a string in the directive code block.

Constructors

GetType
Get the type of an expression via ':type' (or unique prefixes)

GetInfo
Get info about the identifier via ':info' (or unique prefixes)

SetExtension
Enable or disable an extension via ':extension' (or prefixes)

LoadFile
Load a Haskell module.

SetLint
Enable or disable a hlint via ':hlint on' or ':hlint off'

HelpForSet
Provide useful info if people try ':set'.

GetHelp
General help via :? or ':help'.

type LineNumber = Int

type ColumnNumber = Int

type ErrMsg = String

layoutChunks :: String -> [Located String]

Split an input string into chunks based on indentation. A chunk is a line and all lines immediately following that are indented beyond the indentation of the first line. This parses Haskell layout rules properly, and allows using multiline expressions via indentation.

parseDirective

Arguments

:: String
Directive string.

-> Int
Line number at which the directive appears.

-> CodeBlock
Directive code block or a parse error.

Parse a directive of the form :directiveName.

getModuleName :: GhcMonad m => String -> m [String]

Parse a module and return the name declared in the 'module X where' line. That line is required, and if it does not exist, this will error. Names with periods in them are returned piece y piece.

unloc :: Located a -> a

Unlocate something - drop the position.

data Located a

Store locations along with a value.

Constructors

Located LineNumber a