IHaskell.IPython

Description

Documentation

ipythonInstalled :: IO Bool

Check whether IPython is properly installed.

installIPython :: IO ()

Install IPython from source.

updateIPython :: IO ()

Update the IPython source tree and rebuild.

setupIPython :: IO ()

Set up IPython properly.

runConsole :: InitInfo -> IO ()

runNotebook :: InitInfo -> Maybe String -> IO ()

readInitInfo :: IO InitInfo

defaultConfFile :: IO (Maybe String)

getIHaskellDir :: IO String

nbconvert :: ViewFormat -> String -> IO ()

Find a notebook and then convert it into the provided format. Notebooks are searched in the current directory as well as the IHaskell notebook directory (in that order).

data ViewFormat

Constructors

Pdf

Html

Ipynb

Markdown

Latex