IHaskell.Types

Documentation

data Profile

A kernel profile, specifying how the kernel communicates.

Constructors

Profile

Fields

ip :: String

The IP on which to listen.

transport :: String

The transport mechanism.

stdinPort :: Port

The stdin channel port.

controlPort :: Port

The control channel port.

hbPort :: Port

The heartbeat channel port.

shellPort :: Port

The shell command port.

iopubPort :: Port

The Iopub port.

key :: ByteString

The HMAC encryption key.

data Message

A message used to communicate with the IPython frontend.

Constructors

KernelInfoRequest
A request from a frontend for information about the kernel.

Fields

header :: MessageHeader

KernelInfoReply
A response to a KernelInfoRequest.

Fields

header :: MessageHeader

ExecuteRequest
A request from a frontend to execute some code.

Fields

header :: MessageHeader

getCode :: ByteString

The code string.

getSilent :: Bool

Whether this should be silently executed.

getStoreHistory :: Bool

Whether to store this in history.

getAllowStdin :: Bool

Whether this code can use stdin.

getUserVariables :: [ByteString]

Unused.

getUserExpressions :: [ByteString]

Unused.

ExecuteReply
A reply to an execute request.

Fields

header :: MessageHeader

status :: ExecuteReplyStatus

The status of the output.

executionCounter :: Int

The execution count, i.e. which output this is.

PublishStatus

Fields

header :: MessageHeader

executionState :: ExecutionState

The execution state of the kernel.

PublishStream

Fields

header :: MessageHeader

streamType :: StreamType

Which stream to publish to.

streamContent :: String

What to publish.

PublishDisplayData

Fields

header :: MessageHeader

source :: String

The name of the data source.

displayData :: [DisplayData]

A list of data representations.

PublishOutput

Fields

header :: MessageHeader

reprText :: String

Printed output text.

executionCount :: Int

Which input this is.

Which output this is for.

PublishInput

Fields

header :: MessageHeader

inCode :: String

Submitted input code.

executionCount :: Int

Which input this is.

Which output this is for.

data MessageHeader

A message header with some metadata.

Constructors

MessageHeader

Fields

identifiers :: [ByteString]

The identifiers sent with the message.

parentHeader :: Maybe MessageHeader

The parent header, if present.

metadata :: Metadata

A dict of metadata.

messageId :: UUID

A unique message UUID.

sessionId :: UUID

A unique session UUID.

username :: Username

The user who sent this message.

msgType :: MessageType

The message type.

data MessageType

The type of a message, corresponding to IPython message types.

Constructors

KernelInfoReplyMessage

KernelInfoRequestMessage

ExecuteReplyMessage

ExecuteRequestMessage

StatusMessage

StreamMessage

DisplayDataMessage

OutputMessage

InputMessage

type Username = ByteString

A username for the source of a message.

type Metadata = Map ByteString ByteString

A metadata dictionary.

type Port = Int

A TCP port.

replyType :: MessageType -> MessageType

Get the reply message type for a request message type.

data ExecutionState

The execution state of the kernel.

Constructors

Busy

Idle

Starting

data StreamType

Input and output streams.

Constructors

Stdin

Stdout

data MimeType

Possible MIME types for the display data.

Constructors

PlainText

MimeHtml

data DisplayData

Data for display: a string with associated MIME type.

Constructors

Display MimeType String

data ExecuteReplyStatus

Possible statuses in the execution reply messages.

Constructors

Ok

Err

Abort