Main

Description

Documentation

data Args

Constructors

Args IHaskellMode [Argument]

data Argument

Constructors

ServeFrom String
Which directory to serve notebooks from.

Extension String
An extension to load at startup.

ConfFile String
A file with commands to load at startup.

Help
Display help text.

data IHaskellMode

Constructors

None

Notebook

Console

UpdateIPython

Kernel (Maybe String)

View (Maybe ViewFormat) (Maybe String)

main :: IO ()

universalFlags :: [Flag Args]

store :: (String -> Argument) -> String -> Args -> Either String Args

notebook :: Mode Args

console :: Mode Args

kernel :: Mode Args

update :: Mode Args

view :: Mode Args

ihaskellArgs :: Mode Args

noArgs :: Arg a

ihaskell :: Args -> IO ()

addDefaultConfFile :: [Argument] -> IO [Argument]

Add a conf file to the arguments if none exists.

showingHelp :: IHaskellMode -> [Argument] -> IO () -> IO ()

initInfo :: [Argument] -> IO InitInfo

Parse initialization information from the flags.

runKernel

Arguments

:: String
Filename of profile JSON file.

-> InitInfo
Initialization information from the invocation.

-> IO ()

Run the IHaskell language kernel.

initialKernelState :: IO (MVar KernelState)

dupHeader :: MessageHeader -> MessageType -> IO MessageHeader

Duplicate a message header, giving it a new UUID and message type.

createReplyHeader :: MessageHeader -> Interpreter MessageHeader

Create a new message header, given a parent message header.

replyTo :: ZeroMQInterface -> Message -> MessageHeader -> KernelState -> Interpreter (KernelState, Message)

Compute a reply to a message.