IHaskell.Types

Description

Documentation

data Profile

A kernel profile, specifying how the kernel communicates.

Constructors

Profile

Fields

ip :: String

The IP on which to listen.

transport :: String

The transport mechanism.

stdinPort :: Port

The stdin channel port.

controlPort :: Port

The control channel port.

hbPort :: Port

The heartbeat channel port.

shellPort :: Port

The shell command port.

iopubPort :: Port

The Iopub port.

key :: ByteString

The HMAC encryption key.

data Message

A message used to communicate with the IPython frontend.

Constructors

KernelInfoRequest
A request from a frontend for information about the kernel.

Fields

header :: MessageHeader

KernelInfoReply
A response to a KernelInfoRequest.

Fields

header :: MessageHeader

ExecuteRequest
A request from a frontend to execute some code.

Fields

header :: MessageHeader

getCode :: ByteString

The entire block of text where the line is. This may be useful in the case of multiline completions where more context may be needed. Note: if in practice this field proves unnecessary, remove it to lighten the messages. json field block

The code string.

getSilent :: Bool

Whether this should be silently executed.

getStoreHistory :: Bool

Whether to store this in history.

getAllowStdin :: Bool

Whether this code can use stdin.

getUserVariables :: [ByteString]

Unused.

getUserExpressions :: [ByteString]

Unused.

ExecuteReply
A reply to an execute request.

Fields

header :: MessageHeader

status :: ExecuteReplyStatus

The status of the output.

executionCounter :: Int

The execution count, i.e. which output this is.

PublishStatus

Fields

header :: MessageHeader

executionState :: ExecutionState

The execution state of the kernel.

PublishStream

Fields

header :: MessageHeader

streamType :: StreamType

Which stream to publish to.

streamContent :: String

What to publish.

PublishDisplayData

Fields

header :: MessageHeader

source :: String

The name of the data source.

displayData :: [DisplayData]

A list of data representations.

PublishOutput

Fields

header :: MessageHeader

reprText :: String

Printed output text.

executionCount :: Int

Which input this is.

Which output this is for.

PublishInput

Fields

header :: MessageHeader

inCode :: String

Submitted input code.

executionCount :: Int

Which input this is.

Which output this is for.

CompleteRequest

Fields

header :: MessageHeader

getCode :: ByteString

The entire block of text where the line is. This may be useful in the case of multiline completions where more context may be needed. Note: if in practice this field proves unnecessary, remove it to lighten the messages. json field block

The code string.

getCodeLine :: ByteString

just the line with the cursor. json field line

getCursorPos :: Int

position of the cursor (index into the line?). json field cursor_pos

CompleteReply
# The list of all matches to the completion request, such as # ['a.isalnum', 'a.isalpha'] for the above example. matches : list,

# the substring of the matched text # this is typically the common prefix of the matches, # and the text that is already in the block that would be replaced by the full completion. # This would be 'a.is' in the above example. text : str,

# status should be ok unless an exception was raised during the request, # in which case it should be error, along with the usual error message content # in other messages. status : ok }

Fields

header :: MessageHeader

completionMatches :: [ByteString]

completionMatchedText :: ByteString

completionText :: ByteString

completionStatus :: Bool

ObjectInfoRequest

Fields

header :: MessageHeader

objectName :: ByteString

Name of object which was searched for.

Name of object being searched for.

detailLevel :: Int

Level of detail desired (defaults to 0). 0 is equivalent to foo?, 1 is equivalent to foo??.

ObjectInfoReply

Fields

header :: MessageHeader

objectName :: ByteString

Name of object which was searched for.

Name of object being searched for.

objectFound :: Bool

Whether the object was found.

objectTypeString :: ByteString

Object type.

objectDocString :: ByteString

ShutdownRequest

Fields

header :: MessageHeader

restartPending :: Bool

Whether this shutdown precedes a restart.

Whether this shutdown precedes a restart.

ShutdownReply

Fields

header :: MessageHeader

restartPending :: Bool

Whether this shutdown precedes a restart.

Whether this shutdown precedes a restart.

data MessageHeader

A message header with some metadata.

Constructors

MessageHeader

Fields

identifiers :: [ByteString]

The identifiers sent with the message.

parentHeader :: Maybe MessageHeader

The parent header, if present.

metadata :: Metadata

A dict of metadata.

messageId :: UUID

A unique message UUID.

sessionId :: UUID

A unique session UUID.

username :: Username

The user who sent this message.

msgType :: MessageType

The message type.

data MessageType

The type of a message, corresponding to IPython message types.

Constructors

KernelInfoReplyMessage

KernelInfoRequestMessage

ExecuteReplyMessage

ExecuteRequestMessage

StatusMessage

StreamMessage

DisplayDataMessage

OutputMessage

InputMessage

CompleteRequestMessage

CompleteReplyMessage

ObjectInfoRequestMessage

ObjectInfoReplyMessage

ShutdownRequestMessage

ShutdownReplyMessage

type Username = ByteString

A username for the source of a message.

type Metadata = Map ByteString ByteString

A metadata dictionary.

type Port = Int

A TCP port.

replyType :: MessageType -> MessageType

Get the reply message type for a request message type.

data ExecutionState

The execution state of the kernel.

Constructors

Busy

Idle

Starting

data StreamType

Input and output streams.

Constructors

Stdin

Stdout

data MimeType

Possible MIME types for the display data.

Constructors

PlainText

MimeHtml

data DisplayData

Data for display: a string with associated MIME type.

Constructors

Display MimeType String

data ExecuteReplyStatus

Possible statuses in the execution reply messages.

Constructors

Ok

Err

Abort