Main

Description

Documentation

data KernelState

Constructors

KernelState

Fields

getExecutionCounter :: Int

main :: IO ()

kernel

Arguments

:: String
Filename of profile JSON file.

-> IO ()

Run the IHaskell language kernel.

initialKernelState :: IO (MVar KernelState)

dupHeader :: MessageHeader -> MessageType -> Interpreter MessageHeader

Duplicate a message header, giving it a new UUID and message type.

createReplyHeader :: MessageHeader -> Interpreter MessageHeader

Create a new message header, given a parent message header.

replyTo :: ZeroMQInterface -> Message -> MessageHeader -> KernelState -> Interpreter (KernelState, Message)

Compute a reply to a message.